AgenticAssistant/requirements.txt
streamlit>=1.30.0
groq>=0.4.0
langsmith>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
